requirements.txt
bottle==0.13.2
PyYAML==6.0.2
Jinja2==3.1.4
